Add files via upload
overall, we have implemented a project which predicts the sales of big mart.we have satrted with data gathering and afterwords data-proccesing. then we did analysis of data and findout the numerical as well as categorical feature. after that we have done data-preproccesing and model training. we get R2 score of 0.63 and 0.58 respectively on train and test data.